Splitting libspu into individual source files ugh
