+++ config/slus_006.64.yaml
@@ -124,7 +124,34 @@ segments:
       - [0x3CAF0, c, psyq/libetc/video_mode]
       - [0x3CB18, c, psyq/libsn]
       - [0x3CD30, asm]
-      - [0x3CD48, c, psyq/libspu]
+      - [0x3CD48, c, psyq/libspu/SpuInit] # S_I
+      - [0x3CD68, c, psyq/libspu/_SpuInit] # S_INI
+      - [0x3CEDC, c, psyq/libspu/Spu] # SPU
+      - [0x3DA70, c, psyq/libspu/SpuDataCallback] # S_DCB
+      - [0x3DA94, c, psyq/libspu/SpuQuit] # S_Q
+      - [0x3DB10, c, psyq/libspu/SpuInitMalloc] # S_M_INIT
+      - [0x3DB64, c, psyq/libspu/SpuSetNoiseClock] # S_SNC
+      - [0x3DBB4, c, psyq/libspu/SpuSetReverb] # S_SR
+      - [0x3DC84, c, psyq/libspu/_SpuIsInAllocateArea] # S_M_UTIL
+      - [0x3DD90, c, psyq/libspu/SpuReadDecodedData] # S_RDD
+      - [0x3DE00, c, psyq/libspu/SpuSetIRQ] # S_SI
+      - [0x3DF40, c, psyq/libspu/SpuSetIRQCallback] # S_SIC
+      - [0x3DF84, c, psyq/libspu/_SpuCallback] # S_CB
+      - [0x3DFA8, c, psyq/libspu/SpuGetVoiceEnvelopeAttr] # S_GVEA
+      - [0x3E018, c, psyq/libspu/SpuRead] # S_R
+      - [0x3E078, c, psyq/libspu/SpuWrite] # S_W
+      - [0x3E0D8, c, psyq/libspu/SpuSetTransferStartAddr] # S_STSA
+      - [0x3E130, c, psyq/libspu/SpuSetTransferMode] # S_STM
+      - [0x3E164, c, psyq/libspu/SpuSetTransferCallback] # S_STC
+      - [0x3E188, c, psyq/libspu/SpuSetCommonAttr] # S_SCA
+      - [0x3E51C, c, psyq/libspu/SpuSetReverbModeType] # S_SRMT
+      - [0x3E6F8, c, psyq/libspu/_spu_setReverbAttr] # S_SRA
+      - [0x3EBC8, c, psyq/libspu/SpuClearReverbWorkArea] # S_CRWA
+      - [0x3ED64, c, psyq/libapi/WaitEvent] # WaitEvent
+      - [0x3ED74, c, psyq/libspu/SpuSetReverbModeDepth] # S_SRMD
+      - [0x3EDA0, c, psyq/libspu/SpuSetReverbModeDelayTime] # S_SRMDT
+      - [0x3EEB8, c, psyq/libspu/SpuSetReverbModeFeedback] # S_SRMFB
+      - [0x3EF74, c, psyq/libspu/SpuGetReverbModeType] # S_GRMT
       - [0x3EF84, c, psyq/libcard]
       - [0x3F080, asm] # Likely part of libcard
       - [0x3F290, sdata]
