+++ src/slus_006.64/psyq/libapi/WaitEvent.c
@@ -1,0 +1,3 @@
+#include "common.h"
+
+INCLUDE_ASM("asm/slus_006.64/nonmatchings/psyq/libapi/WaitEvent", WaitEvent);
