+++ src/slus_006.64/psyq/libspu/SpuDataCallback.c
@@ -1,0 +1,7 @@
+#include "common.h"
+
+#include "libspu_internal.h"
+
+void _SpuDataCallback(SpuTransferCallbackProc func) {
+    DMACallback(4, func);
+}
