+++ src/slus_006.64/psyq/libspu/SpuGetReverbModeType.c
@@ -1,0 +1,8 @@
+#include "common.h"
+
+#include "libspu_internal.h"
+
+void SpuGetReverbModeType(long* type)
+{
+    *type = _spu_rev_attr.mode;
+}
