+++ src/slus_006.64/psyq/libspu/SpuInit.c
@@ -1,0 +1,7 @@
+#include "common.h"
+
+#include "libspu_internal.h"
+
+void SpuInit(void) {
+    _SpuInit(0);
+}
