+++ src/slus_006.64/psyq/libspu/SpuReadDecodedData.c
@@ -1,0 +1,5 @@
+#include "common.h"
+
+#include "libspu_internal.h"
+
+INCLUDE_ASM("asm/slus_006.64/nonmatchings/psyq/libspu/SpuReadDecodedData", SpuReadDecodedData);
