+++ src/slus_006.64/psyq/libspu/SpuSetTransferCallback.c
@@ -1,0 +1,11 @@
+#include "common.h"
+
+#include "libspu_internal.h"
+
+SpuTransferCallbackProc SpuSetTransferCallback(SpuTransferCallbackProc func) {
+    SpuTransferCallbackProc previousCallback = _spu_transferCallback;
+    if (func != previousCallback) {
+        _spu_transferCallback = func;
+    }
+    return previousCallback;
+}
