+++ src/slus_006.64/psyq/libspu/_SpuCallback.c
@@ -1,0 +1,7 @@
+#include "common.h"
+
+#include "libspu_internal.h"
+
+void _SpuCallback(SpuIRQCallbackProc func) {
+    InterruptCallback(9, func);
+}
